refactor: common `fixtures` to `conftest`
